# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for processing bracketed images with AutoEnhance AI HDR merging. This API handles order creation, image uploads, HDR processing, and real-time status updates via Supabase Realtime.",
        "title": "Instant HDR Backend API",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/orders/{order_id}/process": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Initiates HDR processing and merging of uploaded images using AutoEnhance AI.\n\n**Processing Options:**\n\n**enhance_type** (default: \"property\"):\n- \"property\": Best for real estate - balanced enhancement\n- \"warm\": Warm color grading for cozy feel (AI \u003e= 4.0)\n- \"neutral\": Neutral natural look (AI \u003e= 4.0)\n- \"modern\": Contemporary enhancement (AI \u003e= 4.0)\n\n**sky_replacement** (default: true):\n- Replaces dull skies with attractive blue skies\n\n**cloud_type** (optional):\n- \"CLEAR\": Clear blue sky, \"LOW_CLOUD\": Subtle clouds, \"HIGH_CLOUD\": Dramatic clouds\n\n**window_pull_type** (default: \"WINDOWS_WITH_SKIES\"):\n- \"NONE\": No window enhancement\n- \"ONLY_WINDOWS\": Enhance window views only\n- \"WINDOWS_WITH_SKIES\": Enhance windows + replace exterior skies (AI \u003e= 5.2) - BEST RESULTS\n\n**vertical_correction** (default: true):\n- Straightens tilted walls and vertical lines\n\n**lens_correction** (default: true):\n- Removes wide-angle lens distortion\n\n**upscale** (default: false):\n- AI upscaling to double resolution (increases processing time)\n\n**privacy** (default: false):\n- Blurs faces and license plates\n\n**bracket_grouping** (default: \"by_upload_group\"):\n- \"by_upload_group\": Use groups from upload\n- \"auto\": Sequential sets (every N brackets = 1 HDR, where N = brackets_per_image)\n- \"all\": One mega-HDR from all brackets\n- \"individual\": Separate images (no HDR)\n\n**brackets_per_image** (default: 3, only for \"auto\" mode):\n- How many consecutive brackets to merge into one HDR image\n- Example: 6 brackets + brackets_per_image=3 → 2 HDR images ([1,2,3] and [4,5,6])\n- 3 = Standard HDR, 5 = High dynamic range, 7 = Extreme lighting",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "process"
                ],
                "summary": "Process images with HDR merge",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID (UUID)",
                        "name": "order_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Processing options with defaults shown in model",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.ProcessRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ProcessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.ProcessRequest": {
            "type": "object",
            "properties": {
                "ai_version": {
                    "description": "AIVersion specifies the AI model version to use for processing.\nExamples: \"4.0\", \"5.2\", \"5.x\"\nVersions ending in .x (e.g., \"5.x\") will automatically use the latest minor version.\nDefault: Latest stable version (automatically selected by AutoEnhance)",
                    "type": "string",
                    "example": "5.x"
                },
                "bracket_grouping": {
                    "description": "BracketGrouping specifies how uploaded brackets are organized into HDR images.\nOptions: \"by_upload_group\", \"auto\", \"all\", \"individual\", or custom array\n- \"by_upload_group\" (RECOMMENDED): Groups brackets by group_id assigned during upload\n- \"auto\": Groups brackets sequentially by sets (e.g., every 3 brackets = 1 HDR)\n- \"all\": Merges ALL brackets into ONE HDR image (maximum dynamic range)\n- \"individual\": Each bracket becomes a separate image (no HDR merging)\n- Custom array: [[id1,id2,id3],[id4,id5]] - Specify exact bracket groupings by bracket_id\nDefault: \"by_upload_group\"",
                    "type": "string",
                    "example": "by_upload_group"
                },
                "brackets_per_image": {
                    "description": "BracketsPerImage specifies how many consecutive brackets to group into one HDR image.\nOnly used when bracket_grouping is \"auto\". This tells the system to group brackets sequentially.\n\nExample with 6 brackets and brackets_per_image=3:\n  - Brackets [1,2,3] → HDR Image #1\n  - Brackets [4,5,6] → HDR Image #2\n\nCommon values:\n- 3 (DEFAULT): Standard 3-exposure HDR (underexposed, normal, overexposed)\n  Best for: Most real estate shots, typical bracketing workflows\n- 5: 5-exposure HDR for more dynamic range\n  Best for: High-contrast scenes, sunset/sunrise shots\n- 7: 7-exposure HDR for maximum detail in shadows and highlights\n  Best for: Extreme lighting (bright windows + dark interiors)\n\nNote: More brackets = better HDR but longer processing time\nDefault: 3",
                    "type": "integer",
                    "example": 3
                },
                "cloud_type": {
                    "description": "CloudType specifies what type of clouds to add when sky replacement is enabled.\nOptions: \"CLEAR\", \"LOW_CLOUD\", \"HIGH_CLOUD\"\n- \"CLEAR\": Clear blue sky with minimal clouds\n- \"LOW_CLOUD\": Low-altitude clouds for subtle effect\n- \"HIGH_CLOUD\": High-altitude clouds for more dramatic skies\nDefault: null (AutoEnhance chooses automatically based on scene)",
                    "type": "string",
                    "enum": [
                        "CLEAR",
                        "LOW_CLOUD",
                        "HIGH_CLOUD"
                    ],
                    "example": "CLEAR"
                },
                "enhance_type": {
                    "description": "EnhanceType specifies the type of enhancement to apply to the image.\nOptions: \"property\", \"property_usa\", \"warm\", \"neutral\", \"modern\"\n- \"property\" (DEFAULT): Best for real estate photography - balanced enhancement\n- \"property_usa\": USA-specific real estate enhancement (for AI version \u003c 4.0)\n- \"warm\": Warm color grading for cozy, inviting feel (AI version \u003e= 4.0)\n- \"neutral\": Neutral color grading for natural look (AI version \u003e= 4.0)\n- \"modern\": Modern, contemporary enhancement style (AI version \u003e= 4.0)\nDefault: \"property\"",
                    "type": "string",
                    "enum": [
                        "property",
                        "property_usa",
                        "warm",
                        "neutral",
                        "modern"
                    ],
                    "example": "property"
                },
                "lens_correction": {
                    "description": "LensCorrection removes lens distortion (barrel/pincushion effect).\nCorrects curved lines caused by wide-angle lenses commonly used in real estate.\nRecommended for all real estate photos.\nDefault: true",
                    "type": "boolean",
                    "example": true
                },
                "metadata": {
                    "description": "Optional metadata to store with the processing request for your own tracking",
                    "type": "object",
                    "additionalProperties": true
                },
                "privacy": {
                    "description": "Privacy blurs faces and license plates for privacy compliance.\nUseful for public listings where privacy protection is required.\nDefault: false",
                    "type": "boolean",
                    "example": false
                },
                "sky_replacement": {
                    "description": "SkyReplacement enables AI-powered sky replacement with realistic clouds.\nWhen enabled, dull or overcast skies are replaced with attractive blue skies.\nWorks best with outdoor property photos.\nDefault: true (recommended for real estate)",
                    "type": "boolean",
                    "example": true
                },
                "upscale": {
                    "description": "Upscale increases image resolution using AI upscaling technology.\nDoubles the resolution while maintaining quality.\nWarning: Significantly increases processing time and final file size.\nDefault: false",
                    "type": "boolean",
                    "example": false
                },
                "vertical_correction": {
                    "description": "VerticalCorrection automatically corrects vertical perspective distortion.\nStraightens walls and vertical lines that appear tilted due to camera angle.\nEssential for professional real estate photography.\nDefault: true (highly recommended)",
                    "type": "boolean",
                    "example": true
                },
                "window_pull_type": {
                    "description": "WindowPullType controls how window views are enhanced.\nOptions: \"NONE\", \"ONLY_WINDOWS\", \"WINDOWS_WITH_SKIES\"\n- \"NONE\": No window enhancement (keep original window views)\n- \"ONLY_WINDOWS\": Enhance window views only (bring out detail)\n- \"WINDOWS_WITH_SKIES\" (DEFAULT): Enhance windows AND replace exterior skies visible through windows (requires AI version \u003e= 5.2)\nDefault: \"WINDOWS_WITH_SKIES\" (recommended for best results)",
                    "type": "string",
                    "enum": [
                        "NONE",
                        "ONLY_WINDOWS",
                        "WINDOWS_WITH_SKIES"
                    ],
                    "example": "WINDOWS_WITH_SKIES"
                }
            }
        },
        "models.ProcessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "order_id": {
                    "type": "string"
                },
                "processing_params": {
                    "type": "object",
                    "additionalProperties": true
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}